requirements.txt
discord==2.3.2
discord.py==2.4.0
requests==2.32.2
python-dotenv==0.21.0
